pyproject.toml
[tool.poetry]
name = "tap-googlemeet"
version = "0.0.1"
description = "`tap-googlemeet` is a Singer tap for GoogleMeet, built with the Meltano Singer SDK."
readme = "README.md"
authors = ["Hidde Stokvis"]
keywords = [
    "ELT",
    "GoogleMeet",
]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "<3.12,>=3.7.1"
singer-sdk = { version="^0.30.0" }
fs-s3fs = { version = "^1.1.1", optional = true }
requests = "^2.31.0"
google-api-python-client = "^2.93.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
iso8601 = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
singer-sdk = { version="^0.30.0", extras = ["testing"] }

[tool.poetry.extras]
s3 = ["fs-s3fs"]

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true

[tool.ruff]
ignore = [
    "ANN101",  # missing-type-self
    "ANN102",  # missing-type-cls
]
select = ["ALL"]
src = ["tap_googlemeet"]
target-version = "py37"


[tool.ruff.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.isort]
known-first-party = ["tap_googlemeet"]

[tool.ruff.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-googlemeet = 'tap_googlemeet.tap:TapGoogleMeet.cli'
